requirements.txt
python-docx
markdown
beautifulsoup4
Pillow
requests